refactoring & neue Tests
QVariant::Type:: -> Qvariant::

Dbtable SQL generation Autoinc is always primary key

Dbtable SQL generation foreign key: string constants (ODOU_Actions)

Dbtable SQL generation remove any "type details" to describe table (like unique)

Tdi.writeData -> tdi.insertRecord

Make all table indix fields longlong

ASSERT if unssuported type is used

Get rid of qsl("")

Dbtfield::QSqlField: feed type into base class on construction

QString singleQuote(Qstring s)

Switch to C++17

Refactor DbInsertableString

Replace    xxx == "" with xxx.isEmpty()

Sort header for build on Linux

Add Tablebaseinserter tests

Change random shuffle from Qt to std

Add dbtable / dbfield sql creation tests (sql string compare)

?! d2percent_str sieht falsch aus -> gelöscht, tests & simplere Funktionen eingeführt

+++ DKV2/creditor.h
@@ -28,7 +28,7 @@ struct creditor
     static const dbtable& getTableDef();
     // constructors
     creditor() : ti(getTableDef()) {}
-    creditor (int i) : ti(getTableDef()) { fromDb(i);}
+    creditor (qlonglong i) : ti(getTableDef()) { fromDb(i);}
     // comparison
     bool operator==(const creditor& c) const;
     // setter
@@ -61,7 +61,7 @@ struct creditor
     qlonglong id() const        {return getValue(qsl("id")).toLongLong();}
     void setId(const qlonglong i)     { ti.setValue(qsl("id"), i);}
     // interface
-    bool fromDb(const int id);
+    bool fromDb(const qlonglong id);
     QVariant getValue(const QString& f) const { return ti.getValue(f);}
     QVariant getVariant();
     bool isValid(QString &errortext) const;
